requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
cffi==1.16.0
Django==5.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
sqlparse==0.4.4
typing_extensions==4.9.0
